Cleanse a couple of migrations that cause PostgreSQL problems

+++ db/migrate/20150106232024_attachment_cleanup.rb
@@ -1,7 +1,7 @@
 class AttachmentCleanup < ActiveRecord::Migration
   def change
     # Drop Attachments and Attachment Types tables
-    drop_table :attachments
-    drop_table :attachment_types
+    drop_table :attachments if table_exists? :attachments
+    drop_table :attachment_types if table_exists? :attachment_types
   end
 end
